requirements.txt
colorama==0.4.4
Fiona-1.8.20-cp38-cp38-win_amd64.whl
GDAL-3.3.0-cp38-cp38-win_amd64.whl
numpy==1.20.3
rasterio-1.2.4-cp38-cp38-win_amd64.whl